Added new DObject class methods read, parse and get
In order to clean things up code-wise and reduce the scope of the
somewhat overloaded DObject.new method, these additional class methods
are introduced.
Calling DObject#new with a string will from now on be deprecated, and
removed entirely in a future version.
This is a first commit to adress issue # 26.
Spec examples that use the deprecated method calls have been moved to a
separate spec file, so these can be easily removed in the future.
Added an option to specify whether encoded strings contain the file
meta header or not.