Got rid of the DRead class
Refactored the DICOM read code to get rid of the DRead class and
simplify the code.
In this process, the (deprecated) ability to call DObject#new with a
file argument has been removed, and some logging messages have been
altered.
The DObject#read (instance method) was removed (made a private method of
Parent).