Updated dictionary to the 2011 edition
Updated data elements and unique identifiers dictionary to the 2011
edition of the DICOM standard.
Added a script for creating element and UID dictionaries.
Replaced the Ruby-code dictionary with tab-separated text file
dictionaries, separated in an elements file and a UID file.
Added specs for UIDs and for a couple of 2011 dictionary additions.