Reorganised handling of elements in dictionary
A simple DictionaryElement class was added to hold the values from the
element dictionary, replacing the previous use of arrays holding the VR
and Name information.
This allows for use of other types of information in the future such as
value multiplicity and retired status.
Some modifications made to DLibrary and Parent maintain compatibility.