Fixed a crash issue with the FFFC,FFFC element
It turns out that the Data Set Trailing Padding element is quite tricky.
First of all it is allowed in the standard that this data element is
added to a Sequence, in which case ruby-dicom would crash.

However, properly handling this item-like data element in ruby-dicom
requires a significant effort for a proper implementation, and thus for
the time being a shortcut has been chosen to avoid ruby-dicom from
crashing when reading these tags:

They are simply ignored (removed from the file).