Changed the spec of DLibrary#element(tag)
This method will now always return a DictionaryElement. In cases where
one is not found by hitting the hash of tags, some further checks will
be performed to see if the tag is private, group length, etc.
In such cases a new DictionaryElement will be created and returned.

The name_and_vr method was simplified by using this new element method
instead.