Implemented a gem root directory constant
After the move from a dictionary in ruby-code to an external text file,
loading the dicom library using a relative path suddenly raised an
error with the dictionary not being found.

In order to fix this, the gem root directory constant had to be
implemented. Now the library again succesfully loads either way (by gem
or by relative path).