Added a method for adding a private dictionary
Now it is possible to specify a custom dictionary to be added to
ruby-dicom's dictionary. This can be useful if you've got a known list
of private elements that you would like for ruby-dicom to handle.

The format of the dictionary must follow ruby-dicom's specific tab
separated dictionary format.

This was implemented as DLibrary#add_element_dictionary(file).