Implemented a spec for handling duplicates
Previously ruby-dicom would sliently overwrite any duplicate
sequences/elements that were encountered.

Now the default behavious has been changed to keep the original element
if a duplicate is encountered. In addition, a warning will be logged
when a duplicate is encountered to facilitate troubleshooting.

Lastly, an :overwrite option to DObject#read and #parse has been
implemented which restores the old behaviour of overwriting when
encountering duplicates.