New DLibrary methods for adding UIDs (Fixes #42)
This commit brings feature parity between data elements and uids in
DLibrary. Just as it was possible to add private data elements to the
dictionary, now it is possible to add private UIDs as well.
New DLibrary methods:
#add_element
#add_uid
#add_uid_dictionary

Also refactored code related to the data element dictionaries to
simplify slightly.