Added Anonymizer#anonymize method
This method is a rewrite of the #execute method, and is intended to
replace the #execute method, which has now been deprecated, along with
its related methods.

The refactoring which has gone into the #anonymize method is intended to
clean up and simplify the Anonymizer code base. Insead of separate
methods for handling file folders, a more flexible data argument is
given directly to the #anonymizer method. This argument may be a DICOM
object or a file string, or arrays of both, or directory path(s).

Added a logger_level option to the Anonymizer class which can be used to
select the desired log level of the DObject manipulations during the
anonymization process.