Removed custom encoding handling
Previously we've used custom encoding flags and methods to handle some
corner cases of big endian which seemed to be missing from the Ruby
pack/unpack methods.

However, as of Ruby 1.9.3, all encoding options are available, and we
are now able to get rid of our custom implementation.

This has alot of benefits, such as reducing code complexity and
improving speed.