Moved #add_item from Parent to ElementalParent
A new mix-in module was created for this, which is used by the two
elemental parent classes: Sequence and Item.

Note that this refactor brings a minor spec change:
If trying to call the #add_item method on a DObject, instead of a
general exception, you will now get a NoMethodError.