Refactored the Anonymizer#anonymize method
The #anonymize method, which was introduced in version 0.9.5, was
somewhat overloaded, designed to work on both files, folders or dicom
objects. Unfortunately, it had memory issues, and were not able to
successfully anonymize folders containing thousands of files.

The new #anonymize methods works only on DICOM objects or an array of
DICOM objects.

A new method #anonymize_path has been introduced for anonymizing files
or folders, and it is able to scale without any memory issues and
anonymize thousands of files if needed.